+++ CHANGELOG.md
@@ -1,11 +1,15 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
-and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html). 
 
 ## [Unreleased]
 
 ### Added
- - Barrel: Holds many stacks of a single item.
- - Trash can: 
+ - Barrel: holds many stacks of a single item
+    - Right-click to insert the held stack (if possible)
+    - Double-right-click to insert as many stacks as possible from inventory
+    - Left-click to extract a single item
+    - Crouch-left-click to extract an entire stack